Added Filter Class
Adds Filter class containing a sourceRule, FilterRule, and
DestinationRule objects which define a sorting operation.

End goal of this branch is to define generic filters so that the
Settings class can be designed to hold a list of Filters.

Also intended is to replace the AnimeSorter class with a more generic
Sorter class that uses any Filters from a settings object.