Added FileMetaData interface
Added interface FileMetaData for classes passed between FilterRule and
DestinationRule.

This interface allows for classes such as AnimeEpisode (formerly
FileName) that hold additional data about the file that will be used to
select destination folders. An example is using the extracted name of an
anime series to create a folder for all episodes from the same series,
which is the intended purpose of the DestinationNamedDirectories class.