Added AllRules class, modified Rule factories to accept new rules without needing a list
AllRules class replaces the Enums in RuleSelectorController. This class
will read the downloadsorter.model directory and generate a list of Rule
classes based on the .java files found. There is no longer a need to
register new Rule classes in an Enum.

DestinationFactory, FilterRuleFactory, and SourceFactory all now use
reflection to attempt to generate rules based on their canonical name.
This change allows the factories to  no longer need a list of Rules.
Rules should be instantiated using a constructor that accepts a
String[], or no argument constructor as a fallback.

FileOperation toString() method was modified to correctly write each
Rule to a separate line.

Rule classes have all been updated to call their
.class.getCanonicalName() method as the first value in their toString()
method, this allows them to be read by the new factory methods.